fix: make sidebar panels occupy same height and expand to full height
