+++ src/components/IncidentDetails/IncidentDetails.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { MemoryRouter } from "react-router-dom";
-import { IncidentSeverity } from "../../api/services/incident";
-import { IncidentDetails } from "./index";
+import { IncidentSeverity, IncidentStatus } from "../../api/services/incident";
+import { IncidentSidebar } from "./IncidentSidebar";
 
 export default {
-  title: "IncidentDetails",
-  component: IncidentDetails,
+  title: "IncidentSidebar",
+  component: IncidentSidebar,
   decorators: [
     (Story) => {
       const queryClient = new QueryClient();
@@ -20,24 +20,27 @@ export default {
     }
   ],
   parameters: { actions: { argTypesRegex: "^on.*" } }
-} as ComponentMeta<typeof IncidentDetails>;
+} as ComponentMeta<typeof IncidentSidebar>;
 
 const incident = {
   id: "1aa1a99a-b487-495c-9cd9-a13cf717bc10",
   title: "Test incident 3",
   created_by: "017fb74a-9c90-2074-6b29-61d71507c231",
   commander: {
+    id: "",
+    email: "",
     name: "System",
-    avatar: null
+    avatar: undefined
   },
-  communicator_id: {
+  commander_id: {
     name: "System",
-    avatar: null
+    avatar: undefined
   },
+  communicator_id: "",
   severity: IncidentSeverity.Low,
   description: "test",
   type: "issue",
-  status: "open",
+  status: IncidentStatus.Open,
   created_at: "2022-04-05T07:19:03.546871",
   updated_at: "2022-04-05T07:19:03.546871",
   hypothesis: [
@@ -45,7 +48,7 @@ const incident = {
       id: "401cf26e-e941-4641-b6ab-1860a484b2a2",
       created_by: {
         name: "System",
-        avatar: null
+        avatar: undefined
       },
       incident_id: "1aa1a99a-b487-495c-9cd9-a13cf717bc10",
       parent_id: null,
@@ -56,11 +59,14 @@ const incident = {
       updated_at: "2022-04-05T07:19:04.177198"
     }
   ],
+  parent_id: "",
+  hypotheses: [],
+  involved: [],
   responder: []
 };
 
-const Template: ComponentStory<typeof IncidentDetails> = (arg: any) => (
-  <IncidentDetails {...arg} />
+const Template: ComponentStory<typeof IncidentSidebar> = (arg: any) => (
+  <IncidentSidebar {...arg} />
 );
 
 export const Variant1 = Template.bind({});
